Added support for Apple Touch Icon for iPhone and iTouch
Mobile Safari looks for an icon when you add an icon for a website
to your desktop. This is what tell its where to look:

<link rel="apple-touch-icon" href="/static/apple-touch-icon.png" />

The icon I added is just a screenshot of a zoomed-in version of the
favicon. From what I can find it looks like the icon renders at 57x57,
though some people are saying that it looks better if you actually
make it 158x158. Might be worth is to create a new icon at that size,
though it looks kinda cool all pixelated.

Obviously I wasn't able to test this but hopefully it'll work. If
it doesn't let me know.